lesson 15 kadai, login functions etc.

+++ app/Http/Controllers/TasksController.php
@@ -15,8 +15,17 @@ class TasksController extends Controller
      */
     public function index()
     {
-        $tasks = Task::all();
-        return view('tasks.index', ['tasks' => $tasks,]);
+        $data = [];
+        if(\Auth::check()) {
+            $user = \Auth::user();
+            $tasks = $user->tasks()->orderBy('created_at', 'desc')->paginate(15);
+            $data = [
+                'user' => $user,
+                'tasks' => $tasks
+                ];
+        }
+        
+        return view('tasks.index', $data);
         
     }
 
@@ -46,9 +55,10 @@ class TasksController extends Controller
             ]);
         
         $task = new Task;
-        $task->content = $request->content;
-        $task->status = $request->status;
-        $task->save();
+        $request->user()->tasks()->create([
+            'content' => $request->content,
+            'status' => $request->status,
+            ]);
         
         return redirect('/');
     }
@@ -61,9 +71,13 @@ class TasksController extends Controller
      */
     public function show($id)
     {
-        $task = Task::findOrFail($id);
-        
-        return view('tasks.show', ['task' => $task,]);
+        $task = \App\Task::findOrFail($id);
+        if (\Auth::id() === $task->user_id) {
+            return view('tasks.show', ['task' => $task,]);    
+        } else {
+            return redirect('/');
+        }
+
     }
 
     /**
@@ -74,9 +88,12 @@ class TasksController extends Controller
      */
     public function edit($id)
     {
-        $task = Task::findOrFail($id);
-        
-        return view('tasks.edit', ['task' => $task,]);
+        $task = \App\Task::findOrFail($id);
+        if (\Auth::id() === $task->user_id) {
+            return view('tasks.edit', ['task' => $task,]);    
+        } else {
+            return redirect('/');
+        }
     }
 
     /**
@@ -109,8 +126,11 @@ class TasksController extends Controller
      */
     public function destroy($id)
     {
-        $task = Task::findOrFail($id);
+        $task = \App\Task::findOrFail($id);
+        
+        if (\Auth::id() === $task->user_id) {
         $task->delete();
+        }
         
         return redirect('/');
     }
